testing for GPU
We are testing the GPU code on SSH. To be able to test the CPU code on local, we created a second test file named TestFileWithGPU.cpp, copy pasted from TestFile.cpp with the addition of testing for GPU. To make a distinction between both there are now two makefiles: one for the tests on CPU only, one for all tests.
Notice that in all .cpp and .cu files, because we are compiling them through the makefiles for testing, we commented out the main() functions so we need to uncomment it to compile these files independently. Finally, in the .cu files, i was struggling to compile them through the makefile and AI told me to embbed the functions with << extern "C"  >>